Improve structure, rename component
