+++ pkg/integrations/dash0/dash0.go
@@ -69,7 +69,7 @@ func (d *Dash0) Components() []core.Component {
 	return []core.Component{
 		&QueryPrometheus{},
 		&ListIssues{},
-		&CreateSyntheticCheck{},
+		&CreateHTTPSyntheticCheck{},
 	}
 }
