+++ pkg/integrations/dash0/example_output_create_http_synthetic_check.json
