+++ web_src/src/pages/workflowv2/mappers/dash0/index.ts
@@ -3,13 +3,13 @@ import { canvasesResolveExecutionErrors } from "@/api-client";
 import { withOrganizationHeader } from "@/utils/withOrganizationHeader";
 import { queryPrometheusMapper } from "./query_prometheus";
 import { listIssuesMapper, LIST_ISSUES_STATE_REGISTRY } from "./list_issues";
-import { createSyntheticCheckMapper } from "./create_synthetic_check";
+import { createHttpSyntheticCheckMapper } from "./create_http_synthetic_check";
 import { buildActionStateRegistry } from "../utils";
 
 export const componentMappers: Record<string, ComponentBaseMapper> = {
   queryPrometheus: queryPrometheusMapper,
   listIssues: listIssuesMapper,
-  createSyntheticCheck: createSyntheticCheckMapper,
+  createHttpSyntheticCheck: createHttpSyntheticCheckMapper,
 };
 
 export const triggerRenderers: Record<string, TriggerRenderer> = {};
@@ -17,7 +17,7 @@ export const triggerRenderers: Record<string, TriggerRenderer> = {};
 export const eventStateRegistry: Record<string, EventStateRegistry> = {
   listIssues: LIST_ISSUES_STATE_REGISTRY,
   queryPrometheus: buildActionStateRegistry("queried"),
-  createSyntheticCheck: buildActionStateRegistry("created"),
+  createHttpSyntheticCheck: buildActionStateRegistry("created"),
 };
 
 export async function resolveExecutionErrors(canvasId: string, executionIds: string[]) {
